Initial commit
Move book marks to plain text files. The first line in each
file is the title of the first page. Each page TITLE LINE is
followed by the URL of that page in the next line. Any URL LINE
is followed by either another title line or the end of the file.